+++ core/importhandler/importhandler.py
@@ -210,6 +210,47 @@ class ImportHandler(BaseImportHandler):
                 fp.write('%s\n' % json.dumps(row_data, cls=DecimalEncoder))
         fp.close()
 
+    def store_data_csv(self, output, compress=False):
+        """
+        Stores the given data to file output using CSV format. The output file
+        contains multiple comma-separated rows and header containing names of
+        the columns.
+
+        Keyword arguments:
+        output -- the file to store the data to.
+        compress -- whether we need to archive data using gzip.
+
+        """
+        import csv
+
+        open_mthd = gzip.open if compress else open
+
+        def _encode(data):
+            for key, value in data.iteritems():
+                if isinstance(value, basestring):
+                    data[key] = value.encode('utf-8').replace('\r', '')
+                elif isinstance(value, (dict, list)):
+                    data[key] = json.dumps(value)
+            return data
+
+        with open_mthd(output, 'w') as fp:
+            first_row = next(self)
+            if not first_row:
+                return
+            fieldnames = first_row.keys()
+
+            writer = csv.DictWriter(
+                fp,
+                fieldnames=fieldnames,
+                quotechar="'",
+                quoting=csv.QUOTE_ALL
+            )
+
+            writer.writeheader()
+            writer.writerow(_encode(first_row))
+            for row_data in self:
+                writer.writerow(_encode(row_data))
+
 
 class RequestImportHandler(BaseImportHandler):