+++ changelog.rst
@@ -1,10 +1,18 @@
 Changelog
 =========
 
+Relese 2013-11-17
+-----------------
+* Feature - Added support csv format for train model
+
+Relese 2013-11-11
+-----------------
+* Feature - Added Lda/Lsi vectorizers
+
 
 Relese 2013-06-14
 -----------------
-* Feature -Added compressing when store data to file
+* Feature - Added compressing when store data to file
 
 
 Relese 2013-05-22